+++ src/main/java/io/leangen/graphql/GraphQLSchemaGenerator.java
@@ -196,6 +196,7 @@ public class GraphQLSchemaGenerator {
     private final Map<String, GraphQLDirective> additionalDirectives = new HashMap<>();
     private final List<AnnotatedType> additionalDirectiveTypes = new ArrayList<>();
     private final Map<String, GraphQLType> additionalTypes = new HashMap<>();
+    private final List<Class<?>> additionalJavaTypes = new ArrayList<>();
     private final Set<Comparator<AnnotatedType>> typeComparators = new HashSet<>();
 
     private final String queryRoot;
@@ -732,6 +733,11 @@ public class GraphQLSchemaGenerator {
         return this;
     }
 
+    public GraphQLSchemaGenerator withAdditionalImplementations(Class<?>... additionalTypes) {
+        Collections.addAll(additionalJavaTypes, additionalTypes);
+        return this;
+    }
+
     public GraphQLSchemaGenerator withAdditionalDirectives(Type... additionalDirectives) {
         return withAdditionalDirectives(
                 Arrays.stream(additionalDirectives).map(GenericTypeReflector::annotate).toArray(AnnotatedType[]::new));
@@ -977,8 +983,8 @@ public class GraphQLSchemaGenerator {
                 typeTransformer, basePackages, environment), new TypeMapperRegistry(typeMappers),
                 new SchemaTransformerRegistry(transformers), valueMapperFactory, typeInfoGenerator, messageBundle, interfaceStrategy,
                 scalarStrategy, typeTransformer, abstractInputHandler, new InputFieldBuilderRegistry(inputFieldBuilders),
-                interceptorFactory, directiveBuilder, inclusionStrategy, relayMappingConfig, additionalTypes(), additionalDirectiveTypes,
-                typeComparator, implDiscoveryStrategy);
+                interceptorFactory, directiveBuilder, inclusionStrategy, relayMappingConfig, additionalTypes(),
+                additionalJavaTypes, additionalDirectiveTypes, typeComparator, implDiscoveryStrategy);
         OperationMapper operationMapper = new OperationMapper(buildContext);
 
         GraphQLSchema.Builder builder = GraphQLSchema.newSchema();