+++ src/main/java/io/leangen/graphql/generator/mapping/common/InterfaceMapper.java
@@ -62,8 +62,12 @@ public class InterfaceMapper extends CachingMapper<GraphQLInterfaceType, GraphQL
 
     private void registerImplementations(AnnotatedType javaType, GraphQLInterfaceType type, OperationMapper operationMapper, BuildContext buildContext) {
         if (isImplementationAutoDiscoveryEnabled(javaType)) {
-            buildContext.implDiscoveryStrategy.findImplementations(javaType, getScanPackages(javaType), buildContext).forEach(impl ->
-                    getImplementingType(impl, operationMapper, buildContext)
+            buildContext.implDiscoveryStrategy.findImplementations(javaType, getScanPackages(javaType), buildContext)
+                    .forEach(impl -> getImplementingType(impl, operationMapper, buildContext)
+                            .ifPresent(implType -> buildContext.typeRegistry.registerDiscoveredCovariantType(type.getName(), impl, implType)));
+        } else {
+            getAdditionalImplementations(javaType, buildContext)
+                    .forEach(impl -> getImplementingType(impl, operationMapper, buildContext)
                             .ifPresent(implType -> buildContext.typeRegistry.registerDiscoveredCovariantType(type.getName(), impl, implType)));
         }
     }
@@ -76,6 +80,11 @@ public class InterfaceMapper extends CachingMapper<GraphQLInterfaceType, GraphQL
     @SuppressWarnings("WeakerAccess")
     protected String[] getScanPackages(AnnotatedType javaType) {
         return javaType.isAnnotationPresent(GraphQLInterface.class) ? javaType.getAnnotation(GraphQLInterface.class).scanPackages() : Utils.emptyArray();
+    }
+
+    @SuppressWarnings("WeakerAccess")
+    protected List<AnnotatedType> getAdditionalImplementations(AnnotatedType type, BuildContext buildContext) {
+        return buildContext.additionalImplementationsOf(type);
     }
 
     private Optional<GraphQLObjectType> getImplementingType(AnnotatedType implType, OperationMapper operationMapper, BuildContext buildContext) {