+++ src/main/java/io/leangen/graphql/generator/mapping/common/UnionTypeMapper.java
@@ -25,7 +25,8 @@ public class UnionTypeMapper extends UnionMapper {
     public GraphQLOutputType toGraphQLType(AnnotatedType javaType, OperationMapper operationMapper, Set<Class<? extends TypeMapper>> mappersToSkip, BuildContext buildContext) {
         GraphQLUnion annotation = javaType.getAnnotation(GraphQLUnion.class);
         List<AnnotatedType> possibleJavaTypes = getPossibleJavaTypes(javaType, buildContext);
-        return toGraphQLUnion(annotation.name(), annotation.description(), javaType, possibleJavaTypes, operationMapper, buildContext);
+        final String name = buildContext.typeInfoGenerator.generateTypeName(javaType, buildContext.messageBundle);
+        return toGraphQLUnion(name, annotation.description(), javaType, possibleJavaTypes, operationMapper, buildContext);
     }
 
     @Override
@@ -54,8 +55,16 @@ public class UnionTypeMapper extends UnionMapper {
             possibleTypes = buildContext.implDiscoveryStrategy.findImplementations(javaType, annotation.scanPackages(), buildContext);
         }
         if (possibleTypes.isEmpty()) {
+            possibleTypes = getAdditionalImplementations(javaType, buildContext);
+        }
+        if (possibleTypes.isEmpty()) {
             throw new TypeMappingException("No possible types found for union type " + javaType.getType().getTypeName());
         }
         return possibleTypes;
+    }
+
+    @SuppressWarnings("WeakerAccess")
+    protected List<AnnotatedType> getAdditionalImplementations(AnnotatedType type, BuildContext buildContext) {
+        return buildContext.additionalImplementationsOf(type);
     }
 }