+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Optimize converter performance [#250](https://github.com/leangen/graphql-spqr/issues/250)
 - Use the `_mappedInputField` directive to attach meta data to input fields [#216](https://github.com/leangen/graphql-spqr/issues/216)
 - Generator should support accepting a bean supplier (makes prototype/dependent scoped beans easier to use) [#254](https://github.com/leangen/graphql-spqr/issues/254)
+- Ability to register additional Java types (used as additional interface implementations or possible union types) [#208](https://github.com/leangen/graphql-spqr/issues/208)
 
 ### Changed
 - Jackson and ClassGraph versions upgraded [#226](https://github.com/leangen/graphql-spqr/issues/226) [#227](https://github.com/leangen/graphql-spqr/issues/227)
@@ -22,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Generate GraphQL type names for generic Java types correctly [#255](https://github.com/leangen/graphql-spqr/issues/255)
 - Resolve type references discovered in subscriptions and mutations [#238](https://github.com/leangen/graphql-spqr/issues/238)
 - Duplicate possible types in unions cause `TypeMappingException` and not `IndexOutOfBoundsException` [#256](https://github.com/leangen/graphql-spqr/issues/256)
+- Adding custom directives can cause `NullPointerException` or `IllegalStateException` [#258](https://github.com/leangen/graphql-spqr/issues/258)
 
 ## [0.9.9] - 2018-12-13
 ### Added